cadastro professor

+++ src/br.com.abc/javascore/introducaoclasses/classes/Professor.java
@@ -11,7 +11,7 @@ Crie uma classe de teste para preencher e imprimir os dados desse professor
 public class Professor {
     public String nome;
     public String matricula;
-    public int rg;
-    public int cpf;
+    public long rg;
+    public long cpf;
 
 }
